Backport

+++ src/main/java/com/laytonsmith/abstraction/events/MCEntityRegainHealthEvent.java
@@ -8,8 +8,9 @@ import com.laytonsmith.core.events.BindableEvent;
  * Created by EntryPoint on 2016-12-13.
  */
 public interface MCEntityRegainHealthEvent extends BindableEvent {
-    double getAmount();
-    void setAmount(double amount);
+    int getAmount();
+
+    void setAmount(int amount);
 
     MCEntity getEntity();
